[core] Computed the swing trajectory w.r.t. the base frame
For getting the foothold position, it's convenient to used the
horizontal frame because we can easily deal with the expected ground or
terrain height-map.

Note that is less computational expensive to generate the trajectories
in the base frame. Moreover, it seems to have a more natural, i.e. easy
to track, trajectories. Further testing may be required.