[wip] Moving to quaternion-based dwl
This requires to defined the configuration vector (7D vector for a free-
flyer joint) and tangent vector (6D vector for a free-flyer joint).
Configuration vectors describes the 3D position and quaternion of the
free-flyer joint. Instead the tangent vectors describe the velocity,
acceleration and effort of this joint.

The whole-body state uses the new standard, i.e. quaternion-based. Note
that base_pos is defined using it. Finally, I created a dedicated header
file for Eigen-based typedef.